Modified Login page to inout login and password, since Yandex has only one page.
