+++ src/main/java/com/epam/pf/LoginPage.java
@@ -1,29 +1,42 @@
 package com.epam.pf;
 
+import org.openqa.selenium.Keys;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 
 public class LoginPage extends AbstractPage {
 
-    @FindBy(id = "identifierId") private WebElement loginInput;
-    @FindBy(xpath = "//div[@id='identifierNext']/content/span") private WebElement nextButton;
+    @FindBy(linkText = "Войти") private WebElement goToLoginPage;
+    @FindBy(name = "login") private WebElement loginInput;
+    @FindBy(name = "password") private WebElement pwdInput;
+    @FindBy(xpath = "//button[@type='submit']") private WebElement nextButton;
 
-    private static final String URL = "https:/gmail.com/";
+    private static final String URL = "https://mail.yandex.ru/";
 
     public LoginPage openPage(){
         driver.get(URL);
         return this;
     }
 
-    public LoginPage fillLoginInput(String query){
+    public LoginPage goToLoginPage(){
+        waitForElementVisible(goToLoginPage);
+        return this;
+    }
 
+    public LoginPage enterLogin(String query){
         loginInput.sendKeys(query);
         return this;
     }
 
-    public PasswordPage pressNextButton(){
-        nextButton.click();
-        return new PasswordPage();
+    public LoginPage enterPasswprd(String query){
+        waitForElementVisible(pwdInput);
+        pwdInput.sendKeys(query);
+        return this;
+    }
+
+    public MainMailPage pressNextButton(){
+        nextButton.sendKeys(Keys.ENTER);
+        return new MainMailPage();
     }
 }
